README.hello-global
Tate Gunning is a student-of-life, who developed a new-found passion for fine-art in 2021. This connection expanded quickly with effort and expert vision, into a production of written volumes and an organized structure of photography portfolios. In Phoenix, Arizona, Tate produces for clients across the real estate, cannabis, investment, and project management industries. He is also an avid journalist, tennis player, and an independent thinker who loves telescope gazing and adventures with his Welsh terrier pup, Saint. 

2022 goals include: create & publish more, scheduled daily free-times, build companies, continue to execute, organize and orchestrate.